cli/pyproject.toml
[project]
name = "tablassist"
version = "0.3.0"
description = "AI-assisted table configuration generation for Tablassert — entity resolution, YAML validation, and Biolink documentation lookup."
authors = [
    { name = "Skye Lane Goetz", email = "sgoetz@isbscience.org" },
]
keywords = [
    "knowledge graph",
    "bioinformatics",
    "entity resolution",
    "ncats translator",
    "yaml configuration",
    "table mining",
    "tablassert",
    "tablassist",
    "biolink",
    "data quality control",
]
readme = "README.md"
license = "Apache-2.0"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Framework :: Pydantic",
    "Operating System :: OS Independent",
    "Environment :: Console",
]
requires-python = ">=3.13"
dependencies = [
    "cyclopts>=4.10.1",
    "docling>=2.84.0",
    "fastexcel>=0.19.0",
    "httpx>=0.28.1",
    "lazy-loader>=0.5",
    "openpyxl>=3.1.5",
    "polars>=1.39.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "tablassert>=7.3.2",
    "textract-py3>=3.0.0",
    "trafilatura>=2.0.0",
    "xlsx2csv>=0.8.6",
]

[project.urls]
Homepage = "https://github.com/SkyeAv/Tablassist"
Source = "https://github.com/SkyeAv/Tablassist"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "./src/tablassist",
]

[project.scripts]
tablassist = "tablassist.cli:serve"

[project.optional-dependencies]
rtcompat = [
    "tablassert[rtcompat]>=7.2.1",
    "polars[rtcompat]>=1.39.0",
]
rt = [
    "tablassist[rtcompat]",
]

[tool.uv]
override-dependencies = [
    "six>=1.16.0",
]
environments = [
    "sys_platform != 'win32'",
]

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]

[tool.pytest.ini_options]
testpaths = ["test"]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true
